Port: Add proto container for s2dk Certificate object.
Add a protobuf representation of one of the key objects exposed by the s2dk
library. This object will be treated as an opaque object for library clients
but will provide the foundation for key usage within the s2dk library. In
particular the certificate will be used to provide cryptographic keys for
encryption and verification in most library operations.