Port: "Update opaque MobileCertificateRequest to include version and specify bytes format."
MobileCertificateRequest is a serialization mechanism for transmitting the public key
registered on mobile clients to the CA which will create the final certificate returned
to a client.  This representation is internal to the s2dk library and may be changed at
any time.  Include a version to enable library maintainers to keep track of breaking
changes, and change the name of the underlying bytes field to reflect its expected structure
(pkcs10).