Bump version (#95)
* Change android jvm gradle dependency to api from implementation.

The jvm project contains APIs which should be common across Android
and server Trifle use-cases.  Switching to api() enables them to be
used directly on Android without needing a wrapper.

* Bump version to 0.1.7.

This includes a fix to the android sdk to expose the common jvm api, as well
as the addition of key generation on android.