Create VerifiedData class to expose client data and certs from SignedData.
SignedData Trifle messages are created by using Trifle keys to sign associated
client data and package them with the appropriate certificates. This data needs
to be accessible by clients. Expose it via antoher type which requires
verification as a prerequisite to ensure clients get the protection Trifle is
meant to provide.