Remove jvm library version method. (#107)
This method was never used and instead was intended as a proof-of-concept to get
code sharing working between the original jvm and android packages when Trifle was
first being set up. The Version file is not connected to the actual version of the
library (as represented in Gradle properties) and so is both misleading and a hassle
to update.  The Android portion of Trifle can still use the Version interface to
demonstrate code sharing, although it now uses actual business logic from the jvm
package anyway.